Update code for plugin attribute, replies. Confusing implementation.
Twitter's user timeline api tag for replies is "exclude replies". The
code and documentation represented that backwards where "replies:true"
perceived to include replies in the timeline when the api requires
false to actually include replies.  For the code I changed the
attribute from "replies" to "no_replies" which is more logical
implementation of the attribute.